requirement.txt
openai-whisper
openai==0.28
speechrecognition
pyttsx3
pyaudio
python-decouple